+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/repair/TableRepairJob.java
@@ -15,10 +15,7 @@
 package com.ericsson.bss.cassandra.ecchronos.core.repair;
 
 import java.math.BigInteger;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Iterator;
-import java.util.List;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
 
 import com.ericsson.bss.cassandra.ecchronos.core.JmxProxyFactory;
@@ -51,6 +48,7 @@ public class TableRepairJob extends ScheduledJob
     private final RepairState myRepairState;
     private final RepairConfiguration myRepairConfiguration;
     private final RepairLockType myRepairLockType;
+    private final List<TableRepairPolicy> myRepairPolicies;
 
     private final TableRepairMetrics myTableRepairMetrics;
     private final TableStorageStates myTableStorageStates;
@@ -66,6 +64,7 @@ public class TableRepairJob extends ScheduledJob
         myRepairConfiguration = builder.repairConfiguration;
         myRepairLockType = builder.repairLockType;
         myTableStorageStates = builder.tableStorageStates;
+        myRepairPolicies = builder.repairPolicies;
     }
 
     public TableReference getTableReference()
@@ -99,7 +98,7 @@ public class TableRepairJob extends ScheduledJob
                         replicaRepairGroup, myJmxProxyFactory, myTableRepairMetrics,
                         myRepairLockType.getLockFactory(),
                         new RepairLockFactoryImpl(),
-                        tokensPerRepair));
+                        tokensPerRepair, myRepairPolicies));
             }
 
             return taskList.iterator();
@@ -192,6 +191,7 @@ public class TableRepairJob extends ScheduledJob
         private RepairConfiguration repairConfiguration = RepairConfiguration.DEFAULT;
         private RepairLockType repairLockType;
         private TableStorageStates tableStorageStates;
+        private final List<TableRepairPolicy> repairPolicies = new ArrayList<>();
 
         public Builder withConfiguration(Configuration configuration)
         {
@@ -238,6 +238,12 @@ public class TableRepairJob extends ScheduledJob
         public Builder withTableStorageStates(TableStorageStates tableStorageStates)
         {
             this.tableStorageStates = tableStorageStates;
+            return this;
+        }
+
+        public Builder withRepairPolices(Collection<TableRepairPolicy> tableRepairPolicies)
+        {
+            this.repairPolicies.addAll(tableRepairPolicies);
             return this;
         }