+++ core/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/scheduling/ScheduleManagerImpl.java
@@ -168,6 +168,11 @@ public class ScheduleManagerImpl implements ScheduleManager, Closeable
 
             for (ScheduledTask task : next)
             {
+                if (!validate(next))
+                {
+                    LOG.info("Job {} was stopped, will continue later", next);
+                    break;
+                }
                 hasRun |= tryRunTask(next, task);
             }